Time: 1464 ms (13.81%) | Memory: 536.8 MB (5.08%) - LeetSync

+++ 552-student-attendance-record-ii/student-attendance-record-ii.cpp
@@ -6,7 +6,7 @@ int help(int n,bool Absent,int late,vector<vector<vector<int>>>&dp){
     if(n==0) return 1;
     if(dp[n][late][Absent]!=-1) return dp[n][late][Absent];
 
-    long long a=0,b=0,c=0;
+    long  a=0,b=0,c=0;
     if(!Absent){
         a=help(n-1,true,2,dp);
     }
